create BaseTest.java
