+++ src/main/java/framework/BaseEntity.java
@@ -1,9 +1,6 @@
 package framework;
 
-import framework.utils.WebDriverManager;
 import org.openqa.selenium.WebDriver;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
 
 /**
  * Holds all main classes
@@ -16,20 +13,4 @@ public abstract class BaseEntity {
     //TODO Log field
     protected WebDriver driver;
     private static int step = 1;
-
-    @BeforeClass
-    public void before() {
-        //TODO logger.initStep(step);
-        driver = BrowserFactory.getInstance().getDriver();
-        WebDriverManager.maximize(driver);
-        //TODO Waiter.fluentWait(driver);
-
-
-    }
-
-    @AfterClass
-    public void turnDown() {
-        WebDriverManager.close(driver);
-        //TODO logger.makeSeparator();
-    }
 }
