+++ src/main/java/framework/BaseTest.java
@@ -1,0 +1,21 @@
+package framework;
+
+import framework.utils.WebDriverManager;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+
+public abstract class BaseTest extends BaseEntity {
+    @BeforeClass
+    public void before() {
+        //TODO logger.initStep(step);
+        driver = BrowserFactory.getInstance().getDriver();
+        WebDriverManager.maximize(driver);
+        //TODO Waiter.fluentWait(driver);
+    }
+
+    @AfterClass
+    public void turnDown() {
+        WebDriverManager.close(driver);
+        //TODO logger.makeSeparator();
+    }
+}
